+++ otus/app/Http/Controllers/Favorites/FavoritesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Favorites;
 
 use App\Models\Favorite;
+use App\Policies\Abilities;
 use App\Services\Favorites\FavoriteService;
 use App\Services\Materials\MaterialService;
 use App\Services\Users\UserService;
@@ -22,22 +23,28 @@ class FavoritesController extends Controller {
     }
 
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW_ANY, Favorite::class);
+
         return \view('favorites.list', [
             'favorites' => $this->favoriteService->searchFavorites()
         ]);
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create() {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Favorite::class);
+
         return view('favorites.create', [
             'users' => $this->userService->searchUsers(),
             'materials' => $this->materialService->searchMaterials(),
@@ -45,35 +52,44 @@ class FavoritesController extends Controller {
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::CREATE, Favorite::class);
+
         $this->favoriteService->storeFavorite($request->all());
         return redirect(route('admin.favorites.index'), 301);
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param \App\Models\Favorite $favorite
-     * @return \Illuminate\Http\Response
+     * @param Favorite $favorite
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Favorite $favorite) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::VIEW, $favorite);
+
         return view('favorites.show', [
             'favorite' => $favorite
         ]);
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \App\Models\Favorite $favorite
-     * @return \Illuminate\Http\Response
+     * @param Favorite $favorite
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Favorite $favorite) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $favorite);
+
         return view('favorites.edit', [
             'favorite' => $favorite,
             'users' => $this->userService->searchUsers(),
@@ -82,24 +98,26 @@ class FavoritesController extends Controller {
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Favorite $favorite
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Favorite $favorite
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, Favorite $favorite) {
+
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $this->authorize(Abilities::UPDATE, $favorite);
+
         $this->favoriteService->updateFavorite($favorite, $request->all());
         return redirect(route('admin.favorites.index'), 301);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\Favorite $favorite
-     * @return \Illuminate\Http\Response
+     * @param Favorite $favorite
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Favorite $favorite) {
+        $this->authorize(Abilities::DELETE, $favorite);
         $this->favoriteService->destroyFavorites([$favorite->id]);
     }
 }
